### 2016-09-06 #57 @whedon
[REVIEW]: akutils: Facilitating analyses of microbial communities through QIIME.
**Submitting author:** @alk224 (<a href="http://orcid.org/0000-0002-7057-9957">Andrew Krohn</a>)
**Repository:** <a href="https://github.com/alk224/akutils-v1.2" target ="_blank">https://github.com/alk224/akutils-v1.2</a>
**Version:** v1.2.3

**Reviewers and authors**: Please avoid lengthy details of difficulties in the review thread. Instead, please create a new issue
in the <a href="https://github.com/alk224/akutils-v1.2" target="_blank">target repository</a> and link to those issues (especially acceptance-blockers)
in the review thread below. (For completists: if the target issue tracker is also on GitHub, linking the review thread in the issue or vice
versa will create corresponding breadcrumb trails in the link target.)
