### 2016-09-06 #57 @arfon
Hi @alk224 - thanks for submitting this. Before we can proceed with this submission please take a look at the author guidelines (http://joss.theoj.org/about#author_guidelines) on how to format the `paper.md` file. Specifically, the metadata needs to be formatted _exactly_ as the example shows below. In addition, the references need to be put in a separate bibtex file named `paper.bib` and citations to these entries made using [this formatting](http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citations).

```
---
title: 'Fidgit: An ungodly union of GitHub and figshare'
tags:
  - example
  - tags
  - for the paper
authors:
 - name: Arfon M Smith
   orcid: 0000-0002-3957-2474
   affiliation: GitHub Inc.
date: 14 February 2016
bibliography: paper.bib
---

# Summary

Your summary here....

# References
```
