### 2016-09-06 4fa33a6
fixed bug where update would fail when custom primers were present
