### 2016-09-06 6545e79
fixed bug where update would fail when custom primers were present
