### 2016-09-09 #57 @alk224
Thanks Arfon,

Will try to make the necessary changes this weekend. I appreciate the help.

--
Andrew Krohn
PhD Candidate, NAU Biological Sciences
Research Technician, EnGGen

On Mon, Sep 5, 2016 at 9:50 PM, Arfon Smith <notifications@github.com>
wrote:

> Hi @alk224 <https://github.com/alk224> - thanks for submitting this.
> Before we can proceed with this submission please take a look at the author
> guidelines (http://joss.theoj.org/about#author_guidelines) on how to
> format the paper.md file. Specifically, the metadata needs to be
> formatted *exactly* as the example shows below. In addition, the
> references need to be put in a separate bibtex file named paper.bib and
> citations to these entries made using this formatting
> <http://rmarkdown.rstudio.com/authoring_bibliographies_and_citations.html#citations>
> .
>
> ---
> title: 'Fidgit: An ungodly union of GitHub and figshare'
> tags:
>   - example
>   - tags
>   - for the paper
> authors:
>  - name: Arfon M Smith
>    orcid: 0000-0002-3957-2474
>    affiliation: GitHub Inc.
> date: 14 February 2016
> bibliography: paper.bib
> ---
>
> # Summary
>
> Your summary here....
>
> # References
>
