+++ README.md
@@ -1,7 +1,7 @@
 ###akutils-v1.2 README  
 
 
-This is a collection of scripts I wrote to make myself more productive while doing microbial ecology work with QIIME. I was having difficulty knowing that I was making the best decisions in my analyses, so I started writing this as an exercise in getting to know better the QIIME commands that I was using. I can now use it to run many OTU picking workflows in a row and it generates the plots and stats that I like to have access to right away. This avoids the tedium of entering many commands and helps to keep me organized. I do everything on a Linux system (Ubuntu 14.04). I have also tested most things on CentOS 6.6 (monsoon cluster at NAU). You should be comfortable installing packages and such that may be required. For more information on the functionality of akutils, [check out the wiki pages.](https://github.com/alk224/akutils-v1.2/wiki)  
+This is a collection of scripts I wrote to make myself more productive while doing microbial ecology work with QIIME. I was having difficulty knowing that I was making the best decisions in my analyses, so I started writing this as an exercise in getting to know better the QIIME commands that I was using. I can now use it to run many OTU picking workflows in a row and it generates the plots and stats that I like to have access to right away. This avoids the tedium of entering many commands and helps to keep me organized. I do everything on a Linux system (Ubuntu 14.04). I have also tested most things on CentOS 6.6 (monsoon cluster at Northern Arizona University). You should be comfortable installing packages and such that may be required. For more information on the functionality of akutils, [check out the wiki pages.](https://github.com/alk224/akutils-v1.2/wiki)  
 
 [Also see the home page for short install instructions.](http://alk224.github.io/akutils-v1.2/)  
 
