### 2016-09-18 c907322
updated for JOSS submission
