### 2016-09-18 #57 @alk224
I believe I have fixed the issues raised here. The paper.md should be properly formatted and there is a bibtex file containing the references. I used the tool [here](http://www.snowelm.com/~t/doc/tips/makebib.en.html) to convert to bibtex and cleaned up the output manually. I also used [this tool](https://gist.github.com/arfon/478b2ed49e11f984d6fb) from the JOSS page to produce a metadata file (codemeta.json). Do you require that I also issue a new release that wraps all of these new files into it?
