### 2016-09-19 #57 @arfon
>  Do you require that I also issue a new release that wraps all of these new files into it?

Nope, we don't need those. Thanks for asking.
