### 2016-10-18 dc2dc1a
bug fix for mode 13
