### 2016-10-25 3cba7b8
bug fix in test script at strip primers step
