### 2016-10-25 #57 @alk224
Someone just showed me a bug in the akutils test command. This happened as a result of a recent rewrite of the strip_primers command. The bug is fixed and the test command should function properly now.
