### 2016-10-31 f25114d
fixed logging issue
