### 2016-12-12 #57 @arfon
> Friendly bump on this @tracykteal 😁

👋 @tracykteal - are you still able to complete this review for JOSS?
