### 2017-01-30 bee0590
fixed bug in biom syntax
