### 2017-02-23 #57 @alk224
I haven't gotten any responses from willing individuals. To be clear for whomever does review this, you simply need a clean Ubuntu 14.04 OS installed natively or within a VM to run everything. Anyone new willing to take on this review? Please?
