### 2017-02-26 #57 @serine
Thanks @drpowell, @alk224 yeh I'll be keen to do it. I've done a bit of work on S16 like profiling (viruses instead of bacteria) and used usearch and its related tools, but I also used vsearch faily extensively. My qiime knowledge are more theoretical then practical, never the less I've used most of the tools that qiime wraps around and will be happy to write it up.

p.s @alk224 easy and clean install with miniconda, pip install also works, but user might need to install some tools manually.

p.s.s qiime2 is coming sometime soonish.. don't want to wait for it to come out? I think its mainly migration to python3
