### 2017-02-27 #57 @arfon
> Thanks @drpowell, @alk224 yeh I'll be keen to do it.

Wonderful, thanks @serine. Please take a look at the reviewer guidelines (http://joss.theoj.org/about#reviewer_guidelines) to get started. Basically, you just need to work through the check list above as you proceed. Feel free to ask questions here of the submitting authors.
