### 2017-02-27 #57 @alk224
@serine Thanks so much!

miniconda install came after I started working on this. The installer here: https://github.com/alk224/akutils_ubuntu_installer as per the instructions here: http://alk224.github.io/akutils-v1.2/ will do all of the work for you.

QIIME2 will be released pretty soon, yes, but this is all based on 1.9/1.9.1 so if it doesn't get out soon it will be obsolete. The QIIME team have reached out about setting up an akutils front end for QIIME2, but I just don't have the time right now. I have several data sets that have been processed with this pipeline now so it would be nice to have something to reference directly.
