### 2017-03-07 bcf04a1
bug fix due to change in biom syntax
