### 2017-03-07 c67b4ab
bug fix due to change in biom syntax
