+++ scripts/biomtotxt.sh
@@ -45,7 +45,9 @@ if [[ ! -z "$biom215test" ]]; then
 	biomv="2"
 echo "Using biom version $biomve"
 else
+echo 1
 	biomver=$(biom convert --version >/dev/null)
+echo $biomver
 	biomve=`echo $biomver | cut -d " " -f 4`
 	biomv=`echo $biomve | cut -d "." -f 1`
 echo "Using biom version $biomve"
