### 2017-03-07 e67c6fd
bug fix due to change in biom syntax
