### 2017-03-07 14f339a
bug fix due to change in biom syntax
