### 2017-03-07 4abbbed
bug fix due to change in biom syntax
