### 2017-03-07 f8d6707
bug fix due to change in biom syntax
