### 2017-03-07 0b6353b
bug fix due to change in biom syntax
