### 2017-03-07 374dc2d
bug fix due to change in biom syntax
