+++ scripts/biomtotxt.sh
@@ -113,9 +113,8 @@ Output: $biomname.txt
 		"
 		else
 	informat=$(file $1)
-file $1
-grep "Hierarchical Data Format" $informat
-	informattest=$(grep "Hierarchical Data Format" $informat 2>/dev/null | wc -l)
+file $1 | grep "Hierarchical Data Format" $informat
+	informattest=$(file $1 | grep "Hierarchical Data Format" $informat 2>/dev/null | wc -l)
 	if [[ "$informattest" == "1" ]]; then
 		format="hdf5"
 	else
