+++ scripts/biomtotxt.sh
@@ -112,21 +112,11 @@ Input:  $biomname.biom
 Output: $biomname.txt
 		"
 		else
-	informat=$(file $1)
-file $1 | grep "Hierarchical Data Format" $informat
-	informattest=$(file $1 | grep "Hierarchical Data Format" $informat 2>/dev/null | wc -l)
-	if [[ "$informattest" == "1" ]]; then
-		format="hdf5"
-	else
-		format="json"
-	fi
 		echo "
 There may have been a problem in your conversion.  Check your input and
 try again. If your biom file is hdf5 formatted, you will need to convert
 to json format first.
-
-Input file format: $format
-"
+		"
 
 		fi
 
