### 2017-03-09 #57 @serine
@alk224 Hey sorry took a bit of time to get to this and I also mis-understood what I was reviewing.. I got it :). Your lience missing year and holder

```Copyright (c) <year> <copyright holders>``` bottom of the licence file
