### 2017-03-09 #57 @serine
@alk224 with your git repo. It's a little confusing (to me at least). Usually you'd have a single repo e.g `akutils` and `git clone http://blah/akults.git` and then you have branches and tags (releases) with different versions. I can see that you have at least one other akutils repo - [akutils-v1.1.1](https://github.com/alk224/akutils-v1.1.1). I'm sure you have a reason for that, but little confusing. I guess v1.2 is the next iteraction of v1.1 ?
