### 2017-03-09 #57 @serine
Am I missing something.. but can't get it going...

```
[bioinformatics-platform2]~/downloads/akutils-v1.2 ((v1.2.3))$ bash install

Command autocompletion already configured for akutils.

Command PATH already configured for akutils.

akutils is properly configured. Restart your terminal or run:
source ~/.bashrc

[bioinformatics-platform2]~/downloads/akutils-v1.2 ((v1.2.3))$ akutils test
akutils: command not found
[bioinformatics-platform2]~/downloads/akutils-v1.2 ((v1.2.3))$ source ~/.bashrc
[bioinformatics-platform2]~/downloads/akutils-v1.2 ((v1.2.3))$ akutils test
akutils: command not found
```

I can get it going this way, but then I get stuck at the config step and it just hangs there

```
[bioinformatics-platform2]~/downloads/akutils-v1.2 ((v1.2.3))$ ./scripts/akutils test

No config file is present.  Running config utility to establish configurable
options.

This will help you configure your akutils config file for running akutils workflows.

First, would you like to configure your global settings or make a local config file
to override your global settings?  A local config file will reside within your
current directory.

Or else, you can choose rebuild if you want to generate a fresh global config file.
This is useful if you recently updated akutils and need to integrate newly available
configuration options.

Enter "global," "local," or "rebuild."

local

OK. Checking for existing config file in current directory.
(/home/kirill/downloads/akutils-v1.2/)

No config file detected in local directory. Creating new local config file.
```
