### 2017-03-09 #57 @serine
One other thing before I go to bed.. me personaly I don't like that you "silently" append stuff to `~/.bashrc`.  It's nice thought, but
