### 2017-03-09 #57 @alk224
@serine Sorry about the confusion and the .bashrc. I can add in the license information.

Sometimes the source .bashrc doesn't seem to work and you just have to do a reboot. Also, I've had a few of the sources from the install script miss something this time or that and then just rerun the installer and it usually fills everything in. The .bashrc changes are meant to make it easy to use the akutils commands in a cluster environment where you can call all of the necessary modules rather than having the power to install stuff yourself. I also really enjoy autocomplete (referenced via .bashrc). v1.2 is much better than v1.1 which was my first major coding experience. It is rather clumsily written and lacks command integration and such. I transferred all of the files from v1.1 to 1.2 and rewrote most everything. I'm not good enough at git to really make use of branches, but I should have several releases of v1.2. It would make sense to run
    akutils update
after a reboot to ensure everything is working right. I just had to fix a bug due to a change in the biom command syntax which was breaking a few other commands.
