### 2017-03-09 #57 @alk224
@serine meant to use a code block there. Do:

    akutils update
