### 2017-03-09 #57 @serine
@alk224 sorry but I can't get it working. I took a quick pick at your `intall` script and I couldn't figure out what you are doing with the path.. if I simply `export PATH=/path/to/akutils-v1.2/scripts:$PATH` I can get `akutils` on cmd, but still the same problem - it finds no existing config file, then gives me options to select - I pick `local`.. and nothing happens. I think this is actually a bug. I took a pick at your `scripts/configure_slave.sh` line 120 has an `if` block with only `echo` nothing else happens in this block
