### 2017-03-09 #57 @serine
@arfon Is this intentional to have this question twice?

```
 A statement of need: Do the authors clearly state what problems the software is designed to solve and who the target audience is?
```
