### 2017-03-09 #57 @alk224
@serine Sorry about that. It is entirely likely there is a bug as all the sources seem to shift freely and without announcing that they will do so. I'll run a clean install on a VM in the morning and hopefully get this all working again.
