### 2017-03-09 #57 @serine
I think I missed something..Can akutils work without VM? or can it only work in a particular environmnet? Provided I have all of the tools installed it should work right?
