### 2017-03-09 #57 @arfon
> Its under Documentations and under Software paper

Yes, this is intentional. The point being that there should be a statement of need (presumably the same one?) in both the documentation and the paper.
