### 2017-03-10 #57 @alk224
@serine You want a clean Ubuntu 14.04 environment. See http://alk224.github.io/akutils-v1.2/
