### 2017-03-10 #57 @serine
What does that mean? Like you want no varibales set in your current shell? Can I just get a new shell and unset everything, something like this `env -i env`
