### 2017-03-10 #57 @alk224
@serine Generally just a fresh install using the 14.04.1 .iso. You could use a system that has other things installed, but given all the nuance of a QIIME install, you may find some things present or absent or incorrect versions of dependencies here and there. I have tried to enforce compatible versions during the install which could further break existing tools that you may like and want. I've never run `env -i env` so I'm not sure what that result might be. I know the installer will not work on Ubuntu 12.04 or 16.04 due to differences in package names in the respective repositories and probably similarly with the intermediate versions as well.
