### 2017-03-10 #57 @serine
To me first point is confusing
> 1) Set up your system appropriately. I have done this on a large workstation as well as within a VirtualBox on my netbook. Install Ubuntu 14.04LTS. I have had trouble with 14.04.3 base installs so prefer 14.04.1 install and then update:
http://old-releases.ubuntu.com/releases/14.04.2/ubuntu-14.04.1-desktop-amd64.iso

I think you should either remove it or simplify to something like requires ubuntu 14.04.1 distro. It is still very restrictive. I would have thought that at least within ubuntu everything should be failry similar. Not sure if you want to go this far, but docker or other containers might be great - easy install solutions.

For lets assume my env has been set appropriately and assume I have all dependencies installed

I've done 3 and 4. Now upto running `akutils` step.. and I'm having issues with configuration step..

would you be able to have a look in you `scripts/configure_slave.sh` line ~ 120 (I think in github its 123). There is an `if` block there with one statement `echo blah` and nothing else. I fall into that `if` block leading to `akutils` hanging until I `ctrl^C` ...
