### 2017-03-10 #57 @alk224
@serine I could specify 14.04.1, but that wouldn't strictly be true. My workstation at work is now on 14.04.5 because I continue to run all security updates. Ubuntu repos change with every release, and for some reason as they pack more bits into subsequent iterations of the same release (.1, .2 etc), the system behavior can change here and there. Annoying yes. I do not think I have the time to learn of something like docker given the complexity of the existing installer which I use to restore my workstations (I manage several, though I am a biologist by training) when someone messes them up or the system drive fails.

Curious on the config script. I checked that line of code and there is an open if statement, but it closes later. Look closely and you will see a series of nested conditionals which all close at lines 163, 169 and 170.

Perhaps there is an issue with the config script in that release (or did you clone the repo? That's what I would do). If you downloaded the release instead, try running a `git pull` from within the akutils directory and then try the akutils command with no arguments.
