### 2017-03-10 #57 @serine
My bad, yes `if` statement has bunch of other stuff in it, It was hard to read, cause of differnet indentations style.

I followed installtion guide. I `git clone http:/blah/auktils-v1.2` I tested that, didn't work for me. I then `git checkout v1.2.3` which is your current release right? but still can't get pass configuration step

Any ideas what I am doing wrong?
