### 2017-03-10 #57 @alk224
That was one of my first exercises with conditionals. Some of my inexperience will still be glaring in terms of indents and such.

Can you delete your cloned akutils repo `rm -r ~/akutils-v1.2`

Then try a fresh clone of the repo in your $HOME directory followed by the bash install command from within the cloned directory?
