### 2017-03-10 #57 @serine
```
[bioinformatics-platform2]~$ pwd
/home/kirill
[bioinformatics-platform2]~$ echo $HOME
/home/kirill
[bioinformatics-platform2]~$ git clone https://github.com/alk224/akutils-v1.2.git
Cloning into 'akutils-v1.2'...
remote: Counting objects: 2870, done.
remote: Total 2870 (delta 30), reused 0 (delta 0), pack-reused 2826
Resolving deltas: 100% (1805/1805), done.
Checking connectivity... done.
Checking out files: 100% (342/342), done.
[bioinformatics-platform2]~$ cd akutils-v1.2/
[bioinformatics-platform2]~/akutils-v1.2 (master)$ bash install

Command autocompletion successfully configured for akutils.

Command PATH successfully configured for akutils.

akutils is properly configured. Restart your terminal or run:
source ~/.bashrc

[bioinformatics-platform2]~/akutils-v1.2 (master)$ at
atktopbm  atobm
[bioinformatics-platform2]~/akutils-v1.2 (master)$ akutils test
akutils: command not found
[bioinformatics-platform2]~/akutils-v1.2 (master)$ export PATH=$HOME/akutils-v1.2/scripts:$PATH
[bioinformatics-platform2]~/akutils-v1.2 (master)$ akutils test

No config file is present.  Running config utility to establish configurable
options.

This will help you configure your akutils config file for running akutils workflows.

First, would you like to configure your global settings or make a local config file
to override your global settings?  A local config file will reside within your
current directory.

Or else, you can choose rebuild if you want to generate a fresh global config file.
This is useful if you recently updated akutils and need to integrate newly available
configuration options.

Enter "global," "local," or "rebuild."

local

OK. Checking for existing config file in current directory.
(/home/kirill/akutils-v1.2/)

No config file detected in local directory. Creating new local config file.
```
