### 2017-03-10 #57 @alk224
And did you restart your terminal? I see you entered your own $PATH command.
