### 2017-03-10 #57 @serine
can you refresh, this github page. I removed mine `export` command. Yours worked fine after `source ~/.bashrc`
