### 2017-03-10 #57 @serine
I can run `akutils` but it hangs at the configuration step :)
```
[bioinformatics-platform2]~/akutils-v1.2 (master)$ pwd
/home/kirill/akutils-v1.2
[bioinformatics-platform2]~/akutils-v1.2 (master)$ which akutils
/home/kirill/akutils-v1.2/scripts/akutils
[bioinformatics-platform2]~/akutils-v1.2 (master)$ akutils

No config file is present.  Running config utility to establish configurable
options.

This will help you configure your akutils config file for running akutils workflows.

First, would you like to configure your global settings or make a local config file
to override your global settings?  A local config file will reside within your
current directory.

Or else, you can choose rebuild if you want to generate a fresh global config file.
This is useful if you recently updated akutils and need to integrate newly available
configuration options.

Enter "global," "local," or "rebuild."

local

OK. Checking for existing config file in current directory.
(/home/kirill/akutils-v1.2/)

No config file detected in local directory. Creating new local config file.
```
