### 2017-03-10 #57 @alk224
Looks like there are issues with pip and python. No surprise there in my experience as they change available module names pretty continuously. I know this all worked when I submitted akutils for review back in October. I will get it working again and then ping you on this thread to get the review going again. I don't think I will be able to completely solve these issues today or even this weekend. I really appreciate your willingness to review, and your patience with these issues. If you have used QIIME in the past, you might be accustomed to the rather arduous installs which is what my installer script is supposed to circumvent.
