### 2017-03-12 #57 @serine
Yeh no problem when you ready.

Qiime has miniconda install now. I don't know how long they had it for, probably fairly recent, but it installs everything with ease. Might be worth while considering if you are having issues with pip install. From what I understand conda (miniconda) wraps everything into one "packages" such that there shouldn't be any issues with dependencies.

In general it would be great if you could do few more stdout/stderr messages to keep user upto date with the progress. Like with that config option - I couldn't be sure if `akutils` was doing something or not. It might have been downloading and installing all the `pip` packages, but as far as I was conserned it was just handing there. Don't worry about this for now though.

p.s I'll be on annual next week, so might not hear from that week
