### 2017-08-30 #57 @alk224
Hi Arfon, sorry for the delay. Working on some manuscripts right now so I
can graduate and plan to attend to akutils later this fall.

--
Lela Andrews
PhD Candidate, NAU Biological Sciences
Core Facility Manager, EnGGen

On Wed, Aug 30, 2017 at 10:46 AM, Arfon Smith <notifications@github.com>
wrote:

> 👋 @alk224 <https://github.com/alk224>, how are you getting on?
>
