### 2017-08-30 #57 @arfon
> Hi Arfon, sorry for the delay. Working on some manuscripts right now so I can graduate and plan to attend to akutils later this fall.

OK no problem. I'll remind you again sometime in early October if that works?
