### 2018-02-05 #57 @arfon
If @alk224 doesn't re-appear in the next week I'll assume she's no longer interested in pursing this publication.
