### 2018-02-05 #57 @alk224
Hi @arfon. So sorry about the silence. I'm slowly reaching my goals. Problems with the installer we're identified last week, but until it is fully updated, I still can't address the reviewer comments with respect to silent modifications of .bashrc file. I'm still interested and will advise as soon as this is ready to move back toward review. Thanks for all your patience.
