### 2018-07-19 #57 @alk224
@arfon  thank you for hitting pause on this. Goals are being achieved (grad school completed, manuscripts prepared), but time remains short. Am hoping for time to really update this by December since by now Ubuntu 14.04 will see end of support in less than a year, and QIIME2 is now available with the wonderful provenance feature, but still lacks many options for performing a comprehensive ITS region analysis. As such, this should probably remain paused for now, and I will get back to this as soon as time allows, and simultaneously address reviewer concerns.
