+++ scripts/primer_file.sh
@@ -4,7 +4,7 @@
 #
 #  Version 1.0.0 (June 29, 2016)
 #
-#  Copyright (c) 2014-2016 Andrew Krohn
+#  Copyright (c) 2014-- Lela Andrews
 #
 #  This software is provided 'as-is', without any express or implied
 #  warranty. In no event will the authors be held liable for any damages
