+++ scripts/make_ggplot_taxa_plots.r
@@ -6,7 +6,7 @@
 #
 #  Version 0.0.1 (January 11, 2016)
 #
-#  Copyright (c) 2016 Andrew Krohn
+#  Copyright (c) 2016-- Lela Andrews
 #
 #  This software is provided 'as-is', without any express or implied
 #  warranty. In no event will the authors be held liable for any damages
