### 2018-10-15 136bc4e
fastq_data.sh script
