### 2018-10-15 3060dd2
minor updates
